refactor/#11: useUpdateQuery 커스텀 훅 추출 및 정렬 로직 공통화
